show dmypy errors post serving
After dmypy starts serving, stdout and stderr gets captured.
If we have an error, we assume we can send it to the client.
However, if we have an error outside of client communication, that error
is lost. The easiest way to see this is to run dmypy in daemonize mode,
run a check once, then Control-C to send a KeyboardInterrupt. That
exception is not printed though it should. After this change you can
clearly see it.