Record ignored errors from non-ignored files
Before this change, fine-grained builds could spuriously mark ignored
legitimate errors as "unused ignores".

By keeping track of these ignored errors we ensure that enough analysis
is done to know that the ignored lines are actually useful.

Some other changes are required to keep things working as before because
of the assumption that ignored errors are never recorded.