Record ignored errors from non-ignored files
Before this change, fine-grained builds could spuriously mark ignored
legitimate errors as "unused ignores".

By keeping track of these ignored errors we ensure that enough analysis
is done to know that the ignored lines are actually useful.

We have to change is_errors_for_file so that we don't consider files as
faulty when all of their errors were hidden.