Fixes enum value inference in cases where the value type is of a user-defined data type class (with __new__).
This fixes a regression introduced by #10057 to fix #10000.  The
`not ti.fullname.startswith("builtins.") clause seemed to be intended
to catch enums with a built-in data type like int or bytes, but this is
overly broad.  It should allow any type so long as it is not itself an
enum.Enum subclass.