Support PEP-646 and PEP-692 in the same callable (#16294)
Fixes https://github.com/python/mypy/issues/16285

I was not sure if it is important to support this, but taking into
account the current behavior is a crash, and that implementation is
quite simple, I think we should do this. Using this opportunity I also
improve related error messages a bit.