Properly use proper subtyping for callables (#16343)
Fixes https://github.com/python/mypy/issues/16338

This is kind of a major change, but it is technically correct: we should
not treat `(*args: Any, **kwargs: Any)` special in `is_proper_subtype()`
(only in `is_subtype()`). Unfortunately, this requires an additional
flag for `is_callable_compatible()`, since currently we are passing the
subtype kind information via a callback, which is not applicable to
handling argument kinds.