Allow mypy to output a junit file with per-file results (#16388)
Adds a new `--junit-format` flag to MyPy, which affects the structure of
the junit file written when `--junit-xml` is specified (it has no effect
when not writing a junit file). This flag can take `global` or
`per_file` as values:
* `--junit-format=global` (the default) preserves the existing junit
structure, creating a junit file specifying a single "test" for the
entire mypy run.
* `--junit-format=per_file` will cause the junit file to have one test
entry per file with failures (or a single entry, as in the existing
behavior, in the case when typechecking passes).

In some settings it can be useful to know which files had typechecking
failures (for example, a CI system might want to display failures by
file); while that information can be parsed out of the error messages in
the existing junit files, it's much more convenient to have that
represented in the junit structure.

Tests for the old and new junit structure have been added.