Show attrs auto_detect incompleteness
This test demonstrates that addition of dunder methods in attrs classes
isn't quite complete.

While the existing test demonstrates that it was working for classes
defined with `init=False`, the `__attrs_init__` method wasn't generated
in the same way when `__init__` was defined with `auto_detect=True`
enabled (either explicitly, or by default).

This auto-generated `__attrs_init__` has been enabled in Attrs as of
version 21.1.0.

Related to: https://github.com/python/mypy/issues/10328
Ref: https://github.com/python-attrs/attrs/issues/793
Ref: https://www.attrs.org/en/stable/changelog.html