Document how evil `--no-strict-optional` is (#16731)
On multiple occasions, I've encountered folks using this, running into
issues and then being perplexed when they figure out what
`--no-strict-optional` actually does. Most recently in
https://github.com/python/mypy/issues/16718#issuecomment-1873374058

This is a non-standard, dangerous option that should not be used in
modern typed Python. It's been five and a half years since it was the
default behaviour in mypy, so we should deemphasise and warn about its
existence.