Support type stub generation for `staticmethod` (#14934)
Fixes #13574

This PR fixes the generation of type hints for static methods of
pybind11 classes. The code changes are based on the suggestions in
#13574.

The fix introduces an additional check if the property under inspection
is of type `staticmethod`. If it is, the type information is read from
the staticmethod's `__func__` attribute, instead of the staticmethod
instance itself.

I added a test for C++ classes with static methods bound using pybind11.
Both, an overloaded and a non-overloaded static method are tested.