Improve stubgen tests (#16760)
Improve test cases around #16486

This PR does not change any actual mypy behavior, only hardens the
stubgen tests. The specific changes are:

- **dedicated test cases**: The existing pybind11 test cases originate
from a pybind11 demo. They cover a specific topic involving geometry
types and semi-implemented logic related to it. This somehow distracts
from the aspects we are trying to test here from the mypy stubgen
perspective, because it hides the actual intent of the bindings. I've
simply started adding new test cases that clearly express via their name
what the test case is addressing. I've kept the original demo stuff for
now, so that the new cases are just an addition (overhead is
negligible).
- **running mypy on the generated stubs**: In general it is crucial that
the output produced by the stubgen can actually be type checked by mypy
(this was the regression in #18486). This wasn't covered by the CI check
so far. I've added check now, which would have avoided the regression.
My goal for follow up PRs would be that we can use `mypy
--disallow-untyped-defs` or even `mypy --strict` on the output.
- **minor directory restructuring**: So far the expected stub outputs
were stored in folder names `stubgen` and `stubgen-include-docs`. This
was a bit confusing, because the folder `stubgen` suggested it contains
_the_ stubgen (implementation). I went for `expected_stubs_no_docs` and
`expected_stubs_with_docs` to make the role of the two folders more
explicit.
- **minor script bugfix**: Fix a bug in `test-stubgen.sh`: The
pre-delete functionality was broken, because the `*` was quoted and
therefore did not match.