Stubtest: ignore a new protocol dunder (#16895)
This is added to all protocol classes on Python 3.12.2+ (it was added in
a patch release of 3.12 as part of a bugfix). There's no reason why
you'd want to explicitly include it in a stub (and doing so would lead
the type checker to incorrectly conclude that you wanted a member
literally called `__non_callable_proto_members__`)

Cf. https://github.com/python/typeshed/pull/11384 and
https://github.com/python/typeshed/issues/11383