stubgen: Fix generated dataclass `__init__` signature (#16906)
Fixes #16811

stubgen was swallowing default values for `__init__` methods generated
by the dataclass plugin making their signature incorrect. This is
because the plugin does not include the argument's initializer in the
generated signature. I changed it to include a dummy ellipsis so that
stubgen can generate correct code.
I also fixed arguments added by the dataclass plugin with the invalid
names `*` and `**` to have the valid and unique names `*generated_args`
and `**generated_kwargs` (with extra underscores to make them unique if
necessary). This removes the need for the hack to special case them in
stubgen and is less confusing for someone looking at them in a stub
file.