[mypyc] Optimize away some bool/bit registers (#17022)
If a register is always used in a branch immediately after assignment,
and it isn't used for anything else, we can replace the assignment with
a branch op. This avoids some assignment ops and gotos.

This is not a very interesting optimization in general, but it will help
a lot with tagged integer operations once I refactor them to be
generated in the lowering pass (in follow-up PRs).