[dmypy] sort list of files for update by extension
dmypy receives the list of updated files via `--update` flag.
If this list contains both `foo.py` and `foo.pyi`, the order matters.
It seems to process the first file in the list first. But if we have a
`.pyi` file, we want this to be processed first since this one contains
the typing information.
Let's reverse sort the list of updated files by the extension. This
should be a simple enough fix to resolve this.