Check override compatibility for dynamically typed methods
This fixes #9618 in a fairly aggressive way, by turning on all override
checks for all methods. This means that with this change, MyPy will
complain in cases where just inspecting type signatures involving `Any`
is enough to establish that an override is a type error:
- This includes the case of overriding `@final` as listed in #9618
- But it also includes arity mismatches, for example when an untyped
  subclass method adds a required attribute not in the parent class
  method.

The tests illustrate this, I added a test case to `check-functions.test`
(is this a good place for it?) for the `@final` check specifically but
also updated a few existing tests of dynamic logic.

The resulting code change is very simple (since all override checks
happen in the same place), and I think conceptually it is right to
enable all these checks because the contract is that MyPy won't
generally type check *bodies* of functions, but the function signatures
are actually part of the *class* type definition which is (in other
cases) sanity-checked even when there are no annotations.

If we think that the arity checks are too big a change, I can work
on a refactor that would let us validate *only* `@final` but not other
properties. Maybe waiting for `mypy_primer` output is the way to go?