[mypyc] Inline tagged integer arithmetic and bitwise operations (#17265)
Inline the fast path of various tagged integer operations by using C
inline functions. Most of these operations are very quick, so getting 
rid of the overhead of a C call improves performance significantly. 
This also enables the C compiler to optimize things more, if there 
are constant operands, for example.

This speeds up an older version of the richards benchmark, which didn't
use native integers, by 10% (on CPython 3.12). Even bigger 
improvements are possible in some microbenchmarks.

We didn't do this in the past because of worries about compilation time.
However, I couldn't measure an impact to self-compilation speed, and 
the binary size is only increased by about 0.1%.

Work on mypyc/mypyc#757.