[PEP 695] Fix nested generic classes (#17776)
There was confusion about the fullnames of type variables in nested
generic classes. A type variable could be defined internally as
`m.OuterClass.T`, but it was sometimes accessed as `m.T`. The root cause
was that the semantic analyzer didn't initialize the attribute that
refers to the enclosing class consistently.

Fixes #17596. Fixes #17630.