[PEP 695] Support Annotated[...] in new-style type aliases (#17777)
The rvalue expression isn't semantically analyzed, so we can't rely on
the `fullname` attribute to check if there is a reference to
`Annotated`. Instead, use a lookup function provided by the caller to
determine the fullname.

Error reporting in the second argument to `Annotated` is still
inconsistent, but this seems lower priority. I'll create a follow-up
issue about (or update an existing issue if one exists).

Fixes #17751.