Make "X | Y" union syntax more prominent in documentation
Soon 4 out of 5 supported Python versions will support the `X | Y` syntax,
so we can make it more prominent (Python 3.13 will be out soon, and 3.8
will reach end of life).

The syntax is available starting from Python 3.10, but it can be used
in annotations in earlier Python versions as well if using
`from __future__ import annotations`.