Update various references to deprecated type aliases in docs (#17829)
Use `collections.abc.Iterable`, for example, instead of
`typing.Iterable`, unless we are discussing support for older Python
versions. The latter has been deprecated since Python 3.9. Also update a
few other out-of-date things that I happened to notice.

Part of the motivation is that Python 3.8 will reach end of life in
October, so we can soon start mostly assuming that users are on 3.9 or
newer (even if we'll continue supporting 3.8 for a while still).

---------

Co-authored-by: Jelle Zijlstra <jelle.zijlstra@gmail.com>