Avoid type size explosion when expanding types
If TypedDict A has multiple items that refer to TypedDict B, don't
duplicate the types representing B during type expansion (or generally
when translating types). If TypedDicts are deeply nested, this could
result in lot of redundant type objects.

Example where this could matter (assume B is a big TypedDict):

```
class A(TypedDict):
    a: B
    b: B
    c: B
    ...
    z : B

```

Also deduplicate large unions. It's common to have aliases
that are defined as large unions, and again we want to avoid
duplicating these unions.